docs: how to use html split
