+++ tests/llm_web_kit/pipeline/extractor/html/recognizer/assets/cccode.html
@@ -1,0 +1,9 @@
+<html>
+<span>这是python</span>
+<pre lang="python" class="code-highlight">
+  <cccode>
+    print("Hello, World!")
+  </cccode>
+</pre>
+<span>例子结束</span>
+</html>
